speaker-diarization/gen_subt_v8/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
audio-separator==0.39.1
audioread==3.1.0
av==15.1.0
beartype==0.18.5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
ctranslate2==4.6.1
cycler==0.12.1
Cython==3.2.1
decorator==5.2.1
diffq-fixed==0.2.4
docopt==0.6.2
einops==0.8.1
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
greenlet==3.2.4
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
inputimeout==1.0.4
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.6
lightning-utilities==0.15.2
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.3
nltk==3.9.2
numba==0.62.1
numpy==2.0.2
omegaconf==2.3.0
onnx-weekly==1.21.0.dev20251117
onnx2torch-py313==1.6.0
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
optuna==4.6.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
primePy==1.3
propcache==0.4.1
protobuf==6.33.1
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.23
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pyparsing==3.2.5
pyreadline3==3.5.4
pysubs2==1.8.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==14.2.0
rotary-embedding-torch==0.6.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
samplerate==0.1.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
SQLAlchemy==2.0.44
sympy==1.14.0
tabulate==0.9.0
ten_vad @ git+https://github.com/TEN-framework/ten-vad.git@81fc1791e1a198b4028559504ebf6da6061a4b26
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0+cu126
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0+cu126
torchmetrics==1.8.2
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whisperx==3.7.4
yarl==1.22.0
